backend/requirements.txt
fastapi>=0.115,<0.116
uvicorn[standard]>=0.34,<0.35
websockets>=14,<15
sqlalchemy[asyncio]>=2.0,<2.1
asyncpg>=0.30,<0.31
alembic>=1.14,<1.15
motor>=3.6,<3.7
redis>=5.2,<5.3
pydantic-settings>=2.7,<2.8
bcrypt>=4.2,<4.3
python-jose[cryptography]>=3.3,<3.4
google-generativeai>=0.8,<0.9
openai>=1.60,<1.61
anthropic>=0.42,<0.43
elevenlabs>=1.15,<1.16
mediapipe>=0.10,<0.11
opencv-python-headless>=4.10,<4.11
fer>=22.5,<22.6
ultralytics>=8.3,<8.4
numpy>=1.26,<1.27
requests>=2.32,<2.33
python-dotenv>=1.0,<1.1
structlog>=24.4,<24.5
python-multipart>=0.0.18
httpx>=0.28,<0.29
aiofiles>=24.1,<24.2
email-validator>=2.1,<2.2
